:recycle: Separate curl animation from state
